@@ -19,10 +19,13 @@
         }
     },
     "require": {
-    	"php": ">=5.4.0"
+        "php": "^7.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8|^5.5|^6.5"
+        "doctrine/coding-standard": "^4.0",
+        "phpstan/phpstan": "^0.9.2",
+        "phpstan/phpstan-strict-rules": "^0.9",
+        "phpunit/phpunit": "^5.5|^6.5|^7.0"
     },
     "extra": {
         "branch-alias": {
