@@ -1,52 +1,47 @@
 <?php
 
+declare(strict_types=1);
+
 namespace EasyCSV;
+
+use LogicException;
+use function array_combine;
+use function array_filter;
+use function is_array;
+use function mb_strpos;
+use function sprintf;
+use function str_getcsv;
+use function str_replace;
 
 class Reader extends AbstractBase
 {
-    /**
-     * @var bool
-     */
+    /** @var bool */
     private $headersInFirstRow = true;
 
-    /**
-     * @var array|bool
-     */
+    /** @var string[]|bool */
     private $headers = false;
 
-    /**
-     * @var
-     */
+    /** @var bool */
     private $init;
 
-    /**
-     * @var bool|int
-     */
+    /** @var bool|int */
     private $headerLine = false;
 
-    /**
-     * @var bool|int
-     */
+    /** @var bool|int */
     private $lastLine = false;
 
-    /**
-     * @var bool|int
-     */
+    /** @var bool */
     private $isNeedBOMRemove = true;
 
-    /**
-     * @param $path
-     * @param string $mode
-     * @param bool   $headersInFirstRow
-     */
-    public function __construct($path, $mode = 'r+', $headersInFirstRow = true)
+    public function __construct(string $path, string $mode = 'r+', bool $headersInFirstRow = true)
     {
         parent::__construct($path, $mode);
+
         $this->headersInFirstRow = $headersInFirstRow;
     }
 
     /**
-     * @return bool
+     * @return string[]|bool
      */
     public function getHeaders()
     {
@@ -56,46 +51,46 @@ class Reader extends AbstractBase
     }
 
     /**
-     * @return array|bool
+     * @return string[]|bool
      */
     public function getRow()
     {
         $this->init();
+
         if ($this->isEof()) {
             return false;
         }
 
-        $row = $this->getCurrentRow();
+        $row     = $this->getCurrentRow();
         $isEmpty = $this->rowIsEmpty($row);
 
         if ($this->isEof() === false) {
-            $this->handle->next();
+            $this->getHandle()->next();
         }
 
         if ($isEmpty === false) {
-            return ($this->headers && is_array($this->headers)) ? array_combine($this->headers, $row) : $row;
-        } elseif ($isEmpty) {
+            return ($this->headers !== false && is_array($this->headers)) ? array_combine($this->headers, $row) : $row;
+        }
+
+        if ($isEmpty) {
             // empty row, transparently try the next row
             return $this->getRow();
-        } else {
-            return false;
         }
+
+        return false;
+    }
+
+    public function isEof() : bool
+    {
+        return $this->getHandle()->eof();
     }
 
     /**
-     * @return bool
+     * @return string[][]
      */
-    public function isEof()
+    public function getAll() : array
     {
-        return $this->handle->eof();
-    }
-
-    /**
-     * @return array
-     */
-    public function getAll()
-    {
-        $data = array();
+        $data = [];
         while ($row = $this->getRow()) {
             $data[] = $row;
         }
@@ -103,16 +98,13 @@ class Reader extends AbstractBase
         return $data;
     }
 
-    /**
-     * @return int zero-based index
-     */
-    public function getLineNumber()
+    public function getLineNumber() : int
     {
-        return $this->handle->key();
+        return $this->getHandle()->key();
     }
 
     /**
-     * @return int zero-based index
+     * @return int|bool
      */
     public function getLastLineNumber()
     {
@@ -120,20 +112,21 @@ class Reader extends AbstractBase
             return $this->lastLine;
         }
 
-        $this->handle->seek($this->handle->getSize());
-        $lastLine = $this->handle->key();
+        $this->getHandle()->seek($this->getHandle()->getSize());
+        $lastLine = $this->getHandle()->key();
 
-        $this->handle->rewind();
+        $this->getHandle()->rewind();
 
         return $this->lastLine = $lastLine;
     }
 
     /**
-     * @return array
+     * @return string[]
      */
-    public function getCurrentRow()
+    public function getCurrentRow() : array
     {
-        $current = $this->handle->current();
+        $current = $this->getHandle()->current();
+
         if ($this->isNeedBOMRemove && mb_strpos($current, "\xEF\xBB\xBF", 0, 'utf-8') === 0) {
             $this->isNeedBOMRemove = false;
 
@@ -143,72 +136,79 @@ class Reader extends AbstractBase
         return str_getcsv($current, $this->delimiter, $this->enclosure);
     }
 
-    /**
-     * @param $lineNumber zero-based index
-     */
-    public function advanceTo($lineNumber)
+    public function advanceTo(int $lineNumber) : void
     {
         if ($this->headerLine > $lineNumber) {
-            throw new \LogicException("Line Number $lineNumber is before the header line that was set");
+            throw new LogicException(sprintf(
+                'Line Number %s is before the header line that was set',
+                $lineNumber
+            ));
         } elseif ($this->headerLine === $lineNumber) {
-            throw new \LogicException("Line Number $lineNumber is equal to the header line that was set");
+            throw new LogicException(sprintf(
+                'Line Number %s is equal to the header line that was set',
+                $lineNumber
+            ));
         }
 
         if ($lineNumber > 0) {
-            $this->handle->seek($lineNumber - 1);
+            $this->getHandle()->seek($lineNumber - 1);
         } // check the line before
 
         if ($this->isEof()) {
-            throw new \LogicException("Line Number $lineNumber is past the end of the file");
+            throw new LogicException(sprintf(
+                'Line Number %s is past the end of the file',
+                $lineNumber
+            ));
         }
 
-        $this->handle->seek($lineNumber);
+        $this->getHandle()->seek($lineNumber);
     }
 
-    /**
-     * @param $lineNumber zero-based index
-     */
-    public function setHeaderLine($lineNumber)
+    public function setHeaderLine(int $lineNumber) : bool
     {
-        if ($lineNumber !== 0) {
-            $this->headersInFirstRow = false;
-        } else {
+        if ($lineNumber === 0) {
             return false;
         }
 
+        $this->headersInFirstRow = false;
+
         $this->headerLine = $lineNumber;
 
-        $this->handle->seek($lineNumber);
+        $this->getHandle()->seek($lineNumber);
 
         // get headers
         $this->headers = $this->getRow();
+
+        return true;
     }
 
-    protected function init()
+    protected function init() : void
     {
-        if (true === $this->init) {
+        if ($this->init === true) {
             return;
         }
+
         $this->init = true;
 
-        if ($this->headersInFirstRow === true) {
-            $this->handle->rewind();
-
-            $this->headerLine = 0;
-
-            $this->headers = $this->getRow();
+        if ($this->headersInFirstRow !== true) {
+            return;
         }
+
+        $this->getHandle()->rewind();
+
+        $this->headerLine = 0;
+
+        $this->headers = $this->getRow();
     }
 
     /**
-     * @param $row
-     * @return bool
+     * @param string[]|null[] $row
      */
-    protected function rowIsEmpty($row)
+    protected function rowIsEmpty(array $row) : bool
     {
-        $emptyRow = ($row === array(null));
-        $emptyRowWithDelimiters = (array_filter($row) === array());
-        $isEmpty = false;
+        $emptyRow               = ($row === [null]);
+        $emptyRowWithDelimiters = (array_filter($row) === []);
+        $isEmpty                = false;
 
         if ($emptyRow) {
             $isEmpty = true;
